+++ tSolver/tSolver.h
@@ -3,6 +3,6 @@
 #ifndef C_SOR_3D_TSOLVER_H
 #define C_SOR_3D_TSOLVER_H
 
-void tSolve(const double* f, double* u, N_len Nlen, int iter, double w, double dx);
+void tSolve(const float* f, float* u, N_len Nlen, int iter, float w, float dx);
 
 #endif //C_SOR_3D_TSOLVER_H
